fix: draw background error
