+++ packages/design/src/index.ts
@@ -2,4 +2,3 @@ export type { IClone } from "./IClone";
 export * from "./physics/index";
 export * from "./renderingHardwareInterface/index";
 export * from "./shader-lab/index";
-export * from "./light/index";
