--- packages/design/src/light/index.ts
@@ -1,3 +1,0 @@
-export type { IDirectLightShaderData } from "./IDirectLightShaderData";
-export type { ISpotLightShaderData } from "./ISpotLightShaderData";
-export type { IPointLightShaderData } from "./IPointLightShaderData";
